Renamed full menu bar story files
