+++ .storybook/stories/Examples/FullMenuBar/FullMenuBar.mdx
@@ -1,6 +1,6 @@
 import { Meta, Story, Source, ArgTypes, Subtitle } from '@storybook/blocks';
-import * as Stories from './ExampleMenuBar.stories';
-import source from './ExampleMenuBar.tsx?raw';
+import * as Stories from './FullMenuBar.stories';
+import source from './FullMenuBar.tsx?raw';
 
 <Meta of={Stories} />
